fix: task-8_redux, move locale store
