--- src/actions/actionCreator.js
@@ -1,7 +1,0 @@
-import { CHANGE_LOCALE } from '../constants'
-
-export const changedLocale = (locale) => ({
-  type: CHANGE_LOCALE,
-  locale,
-});
-
