--- src/constants.js
@@ -1,2 +1,0 @@
-export const CHANGE_LOCALE = 'CHANGE_LOCALE'
-export const LANGUAGES = ['En', 'Ru']
